remove form permission provided only to student.

+++ app/api/applies/removeform/[id]/route.ts
@@ -3,6 +3,9 @@ import connectMongo from "@/lib/mongodb";
 import Application from "@/models/Application";
 import { NextRequest,NextResponse } from "next/server";
 import Job from "@/models/Job";
+import Student from "@/models/Student";
+import { authOptions } from "@/lib/authOptions";
+import { getServerSession } from "next-auth";
 
 interface Params{
     params:{id:String}
@@ -14,11 +17,13 @@ export async function DELETE(request:NextRequest,{params}:Params){
         const {id} = await params;
         if(!mongoose.Types.ObjectId.isValid(id.toString())) return NextResponse.json({message:"Invalid job ID"},{status:400});
 
-        //Student Id
-        //This will be obtained from session or cookies 
-        //
-
-        const studentId = "676d61e17501110503524fc6";
+        const session = await getServerSession(authOptions);
+        const studentId = session?.user.id;
+        const Suser = await Student.findById(studentId);
+        if(!Suser){
+            return NextResponse.json({error:"Unauthorized Access"},{status:403});
+        }
+        
         if(!mongoose.Types.ObjectId.isValid(String(studentId).toString())) return NextResponse.json({message:"Invalid student ID"},{status:400});
 
         await Application.findOneAndDelete({student:studentId,job:id})
